# api/models/response.go
package models

// ErrorResponse represents the error structure for API responses
type ErrorResponse struct {
	Error string `json:"error"`
}

// SuccessResponse represents a generic success response
type SuccessResponse struct {
	Message string `json:"message"`
}